pyproject.toml
[tool.poetry]
name = "Riden"
version = "1.1.0"
description = "A python library for Riden RD power supplies"
license = "MIT"
authors = ["Shayne Hartford <shaybox@shaybox.com>"]
homepage = "https://github.com/ShayBox/Riden"
repository = "https://github.com/ShayBox/Riden.git"

[tool.poetry.dependencies]
click = "^8.0.3"
modbus_tk = "^1.1.2"
pyserial = "^3.5"
python = ">=3.7,<4.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[build-system]
requires = ["setuptools", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
riden = "riden.main:main"

[tool.black]
line-length = 80
